made stuff readonly that is not suposed to be changed after initializeation.
stuff that is supposed to be changed in the GUI is made set:able and triggers propertyChanged event.

hiding the implementation details of the Tables and Nodes collections by exposing IEnumerables of those instead of the observerable collection. Will see if this is ok to bind and be updateable in the GUI or if we need to implement INotifyCollectionChanged as well.